+++ src/main/kotlin/org/incava/rando/RandGenList.kt
@@ -1,7 +1,7 @@
 package org.incava.rando
 
 class RandGenList(size: Int, numSlots: Int, numTrials: Int) : RandGen(size, numSlots, numTrials) {
-    private val list: List<Int> = slots.map { it.value.toInt() }
+    private val list: List<Int> = slots.map { it.value }
 
     override fun nextInt(): Int {
         val index = random.nextInt(100)
